split plugin and host compilation targets
