+++ events_test.go
@@ -2,6 +2,7 @@ package vst2_test
 
 import (
 	"fmt"
+	"reflect"
 	"testing"
 
 	"pipelined.dev/audio/vst2"
@@ -28,4 +29,28 @@ func TestEvents(t *testing.T) {
 			fmt.Printf("dump %v\n", string(ev.SysExDump.Bytes()))
 		}
 	}
+}
+
+func assertEqual(t *testing.T, name string, result, expected interface{}) {
+	t.Helper()
+	if !reflect.DeepEqual(expected, result) {
+		t.Fatalf("%v\nresult: \t%T\t%+v \nexpected: \t%T\t%+v", name, result, result, expected, expected)
+	}
+}
+
+func assertNotNil(t *testing.T, name string, result interface{}) {
+	t.Helper()
+	if reflect.DeepEqual(nil, result) {
+		t.Fatalf("%v\nresult: \t%T\t%+v \nexpected: \t%T\t%+v", name, result, result, nil, nil)
+	}
+}
+
+func assertPanic(t *testing.T, fn func()) {
+	t.Helper()
+	defer func() {
+		if r := recover(); r == nil {
+			t.Fatalf("expected panic")
+		}
+	}()
+	fn()
 }
