--- pluginmain/main.go
@@ -1,11 +1,0 @@
-package main
-
-import "pipelined.dev/signal"
-
-var (
-	bufferSize int
-	buf        signal.Floating
-)
-
-func main() {
-}
