+++ vst_linux.go
@@ -1,3 +1,5 @@
+// +build !plugin
+
 package vst2
 
 // #cgo LDFLAGS: -ldl
