+++ vst_linux_test.go
@@ -1,3 +1,5 @@
+// +build !plugin
+
 package vst2_test
 
 import (
