+++ vst_windows.go
@@ -1,3 +1,5 @@
+// +build !plugin
+
 package vst2
 
 import (
